tidy: enable cert checks
Though we don't aim to be certified in any sense, these checks are still
useful. Most of the changes are targeted at cert-err58-cpp, which checks
that no errors are thrown during global initialization and the like.
The check is helpful to us because it points out cases when it would
make sense to use non-allocating globals, like `const char*` or
`string_view`.

The constructors of one function were changed because when the function
names are `const char*`, C++ selects the wrong constructor.
Additionally, the constructor arguments were inconsistently ordered.